+++ i18n/en/common.json
@@ -46,5 +46,9 @@
   },
   "next": "Next",
   "know_more": "Know more",
-  "ago": "ago"
+  "ago": "ago",
+  "signin": "Signin",
+  "subscribe": "My Subscribe",
+  "profile_setting": "Account Settings",
+  "signout": "SignOut"
 }
