+++ i18n/zh/common.json
@@ -46,5 +46,8 @@
   },
   "next": "下一个",
   "know_more": "了解更多",
-  "ago": "前"
+  "ago": "前",
+  "signin": "登录",
+  "profile_setting": "账号设置",
+  "signout": "退出"
 }
